# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智慧物业管理系统 SaaS 多租户版 API
  version: 1.1.1
  description: |
    本接口契约用于约束 Codex 开发。实际开发不得擅自更改路径、核心字段、状态枚举和权限标识。
    所有业务接口默认返回 ApiResponse；列表接口 data 使用 PageResult。
servers:
  - url: https://api.example.com
    description: 生产环境示例
  - url: http://localhost:8080
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
...
  /api/import/batches:
    get:
      tags: [Import]
      summary: 导入批次列表
      x-permission: import:batch:list
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: importType
          in: query
          schema: { type: string }
        - name: importStatus
          in: query
          schema: { $ref: '#/components/schemas/ImportStatus' }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
  /api/import/batches/{batchId}:
    get:
      tags: [Import]
      summary: 导入批次详情
      x-permission: import:batch:view
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/import/batches/{batchId}/errors:
    get:
      tags: [Import]
      summary: 导入错误明细
      x-permission: import:batch:errors
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
  /api/import/batches/{batchId}/errors.csv:
    get:
      tags: [Import]
      summary: 下载导入错误 CSV
      x-permission: import:batch:errors
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: 导入错误 CSV 文件
          content:
            text/csv:
              schema: { type: string, format: binary }
  /api/import/house-csv:
    post:
      tags: [Import, Base]
      summary: 从已上传 CSV 导入房屋
      x-permission: base:house:import
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CsvImportRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/import/history-bill-csv:
    post:
      tags: [Import, Fee]
      summary: 从已上传 CSV 导入历史账单
      x-permission: fee:bill:import
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CsvImportRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/jobs/run-all:
    post:
      tags: [Job]
      summary: 手动执行当前租户或全部租户任务
      x-permission: job:run
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
        - name: leaseExpireDays
          in: query
          schema: { type: integer, default: 30 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/jobs/workorder-sla:
    post:
      tags: [Job]
      summary: 手动执行工单 SLA 超时标记任务
      x-permission: job:workorder:sla
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/jobs/patrol-missed:
    post:
      tags: [Job]
      summary: 手动执行巡检漏检标记任务
      x-permission: job:patrol:missed
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/jobs/lease-expire-remind:
    post:
      tags: [Job]
      summary: 手动执行租赁合同到期提醒任务
      x-permission: job:lease:remind
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/jobs/message-dispatch:
    post:
      tags: [Job]
      summary: 手动执行站内信派发任务
      x-permission: job:message:dispatch
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/resources:
    get:
      tags: [Lease]
      summary: 可租资源列表
      x-permission: lease:resource:list
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string, enum: [VACANT, RESERVED, LEASED, DISABLED] }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
    post:
      tags: [Lease]
      summary: 创建可租资源
      x-permission: lease:resource:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaseResourceRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/customers:
    get:
      tags: [Lease]
      summary: 意向客户列表
      x-permission: lease:customer:list
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string, enum: [POTENTIAL, FOLLOWING, INTENTIONAL, SIGNED, LOST] }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
    post:
      tags: [Lease]
      summary: 创建意向客户
      x-permission: lease:customer:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaseCustomerRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/customers/{customerId}/follows:
    get:
      tags: [Lease]
      summary: 客户跟进记录
      x-permission: lease:customer:follow
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
    post:
      tags: [Lease]
      summary: 新增客户跟进
      x-permission: lease:customer:follow
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaseFollowRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/contracts:
    get:
      tags: [Lease]
      summary: 租赁合同列表
      x-permission: lease:contract:list
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/LeaseContractStatus' }
        - name: customerId
          in: query
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
    post:
      tags: [Lease]
      summary: 创建租赁合同
      x-permission: lease:contract:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaseContractRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/contracts/{contractId}:
    get:
      tags: [Lease]
      summary: 租赁合同详情
      x-permission: lease:contract:view
      parameters:
        - name: contractId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/contracts/{contractId}/activate:
    put:
      tags: [Lease]
      summary: 合同生效并生成租金账单
      x-permission: lease:contract:activate
      parameters:
        - name: contractId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/contracts/{contractId}/terminate:
    put:
      tags: [Lease]
      summary: 终止租赁合同
      x-permission: lease:contract:terminate
      parameters:
        - name: contractId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeaseContractStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/lease/contracts/expire-remind:
    post:
      tags: [Lease]
      summary: 合同到期提醒
      x-permission: lease:contract:remind
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
  /api/app/lease/contracts:
    get:
      tags: [App]
      summary: 我的租赁合同
      x-permission: app:lease:contract:list
      parameters:
        - name: projectId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: mobile
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
...
    BatchId:
      name: batchId
      in: path
      required: true
      schema: { type: integer, format: int64 }
  responses:
    ApiResponse:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    PageResponse:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/PageResult' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true }
        traceId: { type: string }
...
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: { type: object }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
...
    CsvImportRequest:
      type: object
      required: [projectId, sourceFileId]
      properties:
        projectId: { type: integer, format: int64 }
        sourceFileId: { type: integer, format: int64 }
    ImportStatus:
      type: string
      enum: [VALIDATING, VALID_FAILED, IMPORTING, SUCCESS, PARTIAL_SUCCESS, FAILED, VOIDED]
...
    LeaseContractStatus:
      type: string
      enum: [DRAFT, WAIT_EFFECTIVE, ACTIVE, WILL_EXPIRE, EXPIRED, TERMINATED, VOID]
...
    LeaseResourceRequest:
      type: object
      required: [projectId, resourceType, resourceName]
      properties:
        projectId: { type: integer, format: int64 }
        resourceType: { type: string, enum: [HOUSE, SHOP, SPACE, AD, WAREHOUSE] }
        resourceName: { type: string }
        refObjectId: { type: integer, format: int64, nullable: true }
        area: { type: number, format: decimal }
        status: { type: string, enum: [VACANT, RESERVED, LEASED, DISABLED] }
    LeaseCustomerRequest:
      type: object
      required: [projectId, customerName]
      properties:
        projectId: { type: integer, format: int64 }
        customerName: { type: string }
        contactMobile: { type: string }
        sourceChannel: { type: string }
        demandArea: { type: number, format: decimal }
        budgetAmount: { type: number, format: decimal }
        status: { type: string, enum: [POTENTIAL, FOLLOWING, INTENTIONAL, SIGNED, LOST] }
        ownerUserId: { type: integer, format: int64, nullable: true }
    LeaseFollowRequest:
      type: object
      required: [followType, content]
      properties:
        followType: { type: string }
        content: { type: string }
        nextFollowAt: { type: string, format: date-time }
    LeaseContractRequest:
      type: object
      required: [projectId, resourceId, lesseeName, startDate, endDate, rentAmount]
      properties:
        projectId: { type: integer, format: int64 }
        customerId: { type: integer, format: int64, nullable: true }
        resourceId: { type: integer, format: int64 }
        lesseeName: { type: string }
        lesseeMobile: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        rentAmount: { type: number, format: decimal }
        depositAmount: { type: number, format: decimal }
        paymentCycle: { type: string, enum: [MONTH, QUARTER, YEAR] }
        freeRentDays: { type: integer }
        attachmentFileIds: { type: string }
        activeNow: { type: boolean }
    LeaseContractStatusRequest:
      type: object
      properties:
        reason: { type: string }